reorganization and get_changed_patients
