+++ lib/allscripts_api.rb
@@ -5,8 +5,9 @@ require "allscripts_api/configuration"
 require "allscripts_api/utilities/validator"
 require "allscripts_api/magic_params"
 require "allscripts_api/named_magic_methods"
-require "allscripts_api/ordering_methods"
-require "allscripts_api/order"
+require "allscripts_api/orders"
+require "allscripts_api/patients"
+require "allscripts_api/demographics"
 require "allscripts_api/documents"
 require "allscripts_api/client"
 require "allscripts_api/version"
