+++ lib/allscripts_api/demographics.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+require "allscripts_api/demographics/demographics_methods"
