+++ lib/allscripts_api/orders.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+require "allscripts_api/orders/ordering_methods"
+require "allscripts_api/orders/order"
