+++ lib/allscripts_api/patients.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+require "allscripts_api/patients/patient_methods"
